+++ CHANGELOG.md
@@ -11,6 +11,8 @@ Added a new class of tasks that execute after the file has been successfully upl
  * Added hungRunPollingInterval property (can be set in validatorConfig.json) which controls the frequency (6 hours by default) that the database is queried for hung jobs (jobs which are still marked as running after the runMaximumDuration has passed). Found hung jobs are immediately terminated. (#479)
  * Added additional console reports on edit operations. (#507)
  * Jobs are allowed to complete (or timeout) when the server receives a SIGTERM signal. (#512)
+ * A note was added above the Edit Validation "Save" button when any of the input fields fails validation. (#519)
+ * Modified how responses from external processes are received. Should avoid incorrect parsing of JSON responses. (#529)
 
 ## 1.0.9
 ### New Features